Testes backend

+++ backend/src/controllers/IncidentController.js
@@ -41,12 +41,17 @@ module.exports = {
             .select('ong_id')
             .first();
 
-        if (incident.ong_id !== ong_id) {
-            return res.status(401).json({ error: 'Operation not permitted.' })
+        if (incident) {
+            if (incident.ong_id !== ong_id) {
+                return res.status(401).json({ error: 'Operation not permitted.' })
+            }
+
+            await connection('incidents').where('id', id).delete();
+
+            return res.status(204).send();
+        } else {
+            res.status(404).send();
         }
 
-        await connection('incidents').where('id', id).delete();
-
-        return res.status(204).send();
     }
 }
